A lot of stuff added.
1. element wise power
2. Print matrix to a file
3. MATLAB colon operator added as stepspace.
4.  '*' operator improved to handle matrix multiplication with a scalar
5.  Started working on ode45.

+++ cu_mat_lib/block_dim.cu
@@ -1,4 +1,4 @@
-size_t block_dim(int n_ele)
+size_t block_dim(size_t n_ele)
 {
 	size_t tc;
 	if (n_ele<=1024) return(n_ele);
